Added google maps connection to find distances between businesses

+++ src/main/java/com/cs201/barcrawl/repository/BusinessRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface BusinessRepository extends CrudRepository<Business, Integer> {
+public interface BusinessRepository extends CrudRepository<Business, Long> {
     // Nothing to see here~
 }
